fix: GetMealFoodResponse.FoodInfo kcal 값 double->int

+++ src/main/java/mealplanb/server/dto/meal/GetMealFoodResponse.java
@@ -23,6 +23,6 @@ public class GetMealFoodResponse {
         private long foodId;
         private int quantity;
         private String name;
-        private double kcal;
+        private int kcal;
     }
 }
